Ajustes de acordo com requisitos do projeto da matéria Desenvolvimento Web

+++ src/main/java/br/senac/seguradora/projetoseguradora/service/ClienteService.java
@@ -83,7 +83,14 @@ public class ClienteService {
 	}
 
 	public boolean excluirPorId(Integer id) {
-		clienteRepository.deleteById(id);
+		boolean seguroVigente = verificarSeguroVigente();
+		if(!seguroVigente) {
+			clienteRepository.deleteById(id);
+		}
+		return seguroVigente;
+	}
+
+	private boolean verificarSeguroVigente() {
 		return true;
 	}
